Major Updates -----
- dev/test/prod ENV
- nodes can be build from vanilla or with extras
- extras as own build (all or nothing)
- breakup compose to simplify data workers
- build node script
- breakup build for each object (move to compose?)
- clean-up

TODO:
-----

- spark install
- pig/hive property setup
- examples w/ pig, hive and spark
- readme and documentation

+++ scripts/add_data_hadoop.sh
@@ -4,14 +4,18 @@ datafile=$1
 
 function addCSVDataHadoop() {
 
-  docker exec namenode hdfs dfs -ls /data/$fname;
-  docker exec namenode hdfs dfs -mkdir -p /data/
-  fname="$(echo "$datafile" | cut -d'.' -f1).csv"
-  docker cp "$datafile" namenode:/"$datafile"
-  docker exec namenode unzip "$datafile"
-  docker exec namenode hadoop fs -put -f "$fname" /data
-  docker exec namenode rm "$datafile" "$fname"
+    fname="$(echo "$datafile" | cut -d'.' -f1).csv"
+    # check if file on hdfs
+    docker exec namenode hdfs dfs -test -f /data/$fname;
+    # if no, load it
+    if [ $? -eq 1 ]; then
+        docker exec namenode hdfs dfs -mkdir -p /data/
+        docker cp "$datafile" namenode:/"$datafile"
+        docker exec namenode unzip "$datafile"
+        docker exec namenode hadoop fs -put -f "$fname" /data
+        docker exec namenode rm "$datafile" "$fname"
+    fi
 
 }
-
+export -f addCSVDataHadoop
 addCSVDataHadoop
